requirements.txt
httplib2==0.9.2
requests==2.10.0
telepot==8.3
google_api_python_client==1.5.1
python_dateutil==2.5.3
facebook==0.0
feedparser==5.2.1
